+++ packages/browser/src/index.ts
@@ -5,3 +5,4 @@ export * from "./core/cache/TipCache";
 export * from "./core/cache/memoize";
 export * from "./core/json";
 export * from "./core/network";
+export * from "./async";